unfurl optimizations
1.
unfortunately most of our index endpoints support pagination and since
they're GET endpoints the page+per values must be transferred via url
params. however, we also do have POST (/search) endpoints that expect
the same page+per values to be part of the body. chipmunks 'action'
ignores the body when not applicable so it's safe to set it all the time
for params + body.

2.
run all the paginated requests in parallel to speedup things