fixes resolving has many associations via anyOf
a while ago, we've changed all references to always either allow the
association to point to an actual record or not to exist, i.e. being
'null'.
however, the code to resolve association data was not updated to reflect
this.
this only pops up now, since we haven't been using _has many_
associated records pointing to mm3 records yet.

however for the new asset structure an asset is linking to multiple
- artefacts and
- operations